Purchase use case -  no Test

+++ src/java/test/Bridge/Proxy.java
@@ -94,7 +94,7 @@ public class Proxy implements IService {
     }
 
     @Override
-    public Response buyShoppingCart(String visitorName, int expectedPrice, PaymentMethod paymentMethod, Address address) {
+    public Response buyShoppingCart(String visitorName, double expectedPrice, PaymentMethod paymentMethod, Address address) {
         return null;
     }
